% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frLib.R
\docType{package}
\name{frLib-package}
\alias{frLib}
\alias{frLib-package}
\title{Accessory R library}
\description{
An R package containing some useful accessory R functions
}
\details{
This package is currentlydeveloped to run analysis of IgSeq data. 
	
		Please refer to the vignettes to see details.
}
\references{
Feng, et al 2019 \url{https://doi.org/10.1101/483800}
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
